# Slight optimization for std.conv.text/wtext/dtext