@JackStouffer 2017-01-13
Auto-merge toggled on