@ibuclaw 2017-08-03
This caused a regression.

https://issues.dlang.org/show_bug.cgi?id=17712