@BBasile 2018-06-11
This change make things working again.

```diff
@@ -4250,11 +4250,11 @@ private S textImpl(S, U...)(U args)
         {
             static if (
                 is(Unqual!(typeof(arg)) == uint) || is(Unqual!(typeof(arg)) == ulong) ||
                 is(Unqual!(typeof(arg)) == int) || is(Unqual!(typeof(arg)) == long)
             )
-                app.put(arg.toChars);
+                app.put(textImpl!(S)(arg));
             else
                 app.put(to!S(arg));
         }
 
         return app.data;
```
Maybe some templates got discarded because of variadic arg here ?