@BBasile 2018-06-11
That maybe more simply because toChars is not called anymore.